requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document Processing
pypdf2==3.0.1
PyPDF2==3.0.1
pymupdf==1.23.8
python-docx==1.1.0

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10

# LLM Integration
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
numpy==1.26.2
pandas==2.1.4
aiofiles==23.2.1
